move subclasses into own file
